Move widgets to top-level Components per Phlex style guide
`.claude/phlex_style_guide.md` is explicit:

> Favor Top-Level Components Over Namespacing
>
> Prefer flat, top-level component organization over nested
> namespaces. Reason: The phlex-rails Kit syntax (which enables
> automatic rendering of components by name) does not work with
> namespaced components.

The existing project components in this codebase consistently honor
the rule (11 top-level Components::Project*: ProjectViolationsForm,
ProjectBanner, ProjectForm, ProjectAliases, etc.). My initial pass
followed the 3 outliers under Components::Projects::* (Tabs,
LocationsTable, ObsFooter — likely older code); rename to match the
documented convention.

  Components::Projects::TargetNamesWidget
    → Components::ProjectTargetNamesWidget
  Components::Projects::TargetLocationsWidget
    → Components::ProjectTargetLocationsWidget

Files moved with `git mv` to preserve history:
  app/components/projects/target_names_widget.rb
    → app/components/project_target_names_widget.rb
  app/components/projects/target_locations_widget.rb
    → app/components/project_target_locations_widget.rb
  test/components/projects/target_names_widget_test.rb
    → test/components/project_target_names_widget_test.rb
  test/components/projects/target_locations_widget_test.rb
    → test/components/project_target_locations_widget_test.rb

Class declarations and the four call sites (checklists/show.rb,
target_names/_checklist_update.erb,
target_locations/_locations_update.erb,
locations/index.html.erb) updated to the new top-level names, plus
two stale references in controller comments.

Cleaning up the existing `Components::Projects::*` outliers is out
of scope for this PR.
