+++ app/components/project_target_names_widget.rb
@@ -1,0 +1,41 @@
+# frozen_string_literal: true
+
+# Inline form for adding target names to a Project. Posts a list
+# of names (textarea, newline-separated) to
+# `project_target_names_path` via Turbo, so the surrounding
+# checklist re-renders without a full page reload.
+#
+# Pattern B: creates its own FormObject internally so the view
+# only needs to pass the `project:` kwarg.
+class Components::ProjectTargetNamesWidget < Components::ApplicationForm
+  # Optional positional model arg is accepted for ModalForm
+  # compatibility (ignored) — see Pattern B in
+  # .claude/phlex_style_guide.md.
+  def initialize(_model = nil, project:, **)
+    @project = project
+    super(FormObject::ProjectTargetNamesAdd.new,
+          id: "target_names_widget", local: false, **)
+  end
+
+  def around_template
+    @attributes[:class] = "form-inline mb-3"
+    super
+  end
+
+  def view_template
+    super do
+      autocompleter_field(
+        :names,
+        type: :name,
+        textarea: true,
+        separator: "\n",
+        label: :project_target_names_to_add_label.t
+      )
+      submit(:project_target_name_add.t, class: "ml-2 mt-2")
+    end
+  end
+
+  def form_action
+    project_target_names_path(project_id: @project.id)
+  end
+end
