+++ app/controllers/projects/target_locations_controller.rb
@@ -61,7 +61,7 @@ module Projects
     end
 
     # Submitted under the `project_target_locations_add` namespace by
-    # Components::Projects::TargetLocationsWidget (Pattern B form).
+    # Components::ProjectTargetLocationsWidget (Pattern B form).
     def parse_locations_from_params
       raw = params.dig(:project_target_locations_add, :locations).to_s
       raw.split("\n").filter_map do |entry|
