+++ app/views/controllers/projects/locations/index.html.erb
@@ -4,7 +4,7 @@ add_project_banner(@project)
 %>
 
 <% if @project.is_admin?(@user) %>
-  <%= render(Components::Projects::TargetLocationsWidget.new(
+  <%= render(Components::ProjectTargetLocationsWidget.new(
                project: @project
              )) %>
 <% end %>
