+++ app/views/controllers/projects/target_names/_checklist_update.erb
@@ -1,5 +1,5 @@
 <%= turbo_stream.replace("target_names_widget") do %>
-  <%= render(Components::Projects::TargetNamesWidget.new(project: project)) %>
+  <%= render(Components::ProjectTargetNamesWidget.new(project: project)) %>
 <% end %>
 <%
 # Carry any optional scope fields the caller passed through to the
